Fix Create Event & Template Event Details Screen
Context: 
When you created an event or copied one over from a template, and tried to view it from Your Events tab, the menu courses and items were not showing up. 

Important To Note
1. When you created an event from scratch and it populated in your "Your Events" tab, we would try to grab the menu from the menu templates or the chef's menu's collection. Keep in this in mind as I explain point 2. 
2. When you copied over an event template, we would point to the menu template id and try to grab said menu in the menu templates collection. 

Why this was an issue: 
In either option above, we can can add a flag for the pageName and get our menu courses and items without an issue. 

The problem lies when a copied template is viewed from Your Events because now we don't know where the menu came from originally. 

What I did was:
- Add a flag in the experiences collection to know if the event is[a]Template.
- You can see the menus now on each event when you press on each one in the templates tab.
- Added a flag for only the pricing to show up in the templates if you navigate to the templates tab in both menu and events. 
- Added a tooltip so the user wouldn't be confused about what they see.